fix(can/gs_usb): convert timestamp_us from u32 to u64
GS-USB frame uses u32 for timestamp_us, but PiperFrame requires u64.
This conversion ensures type compatibility and consistency with the
state layer design (JointPositionState.hardware_timestamp_us: u64).

+++ src/can/gs_usb/mod.rs
@@ -295,7 +295,7 @@ impl CanAdapter for GsUsbCanAdapter {
                         data: gs_frame.data,
                         len: gs_frame.can_dlc.min(8),
                         is_extended: (gs_frame.can_id & CAN_EFF_FLAG) != 0,
-                        timestamp_us: gs_frame.timestamp_us,
+                        timestamp_us: gs_frame.timestamp_us as u64, // GS-USB 使用 u32，转换为 u64
                     };
                     trace!("Received CAN frame: ID=0x{:X}, len={}", frame.id, frame.len);
                     return Ok(frame);
